+++ src/lib.rs
@@ -2,6 +2,7 @@
 #![allow(clippy::too_many_arguments)]
 //requires nightly, or later stable version
 //#![warn(clippy::unwrap_used)]
+#![allow(clippy::upper_case_acronyms)]
 
 pub mod post;
 pub mod seal;
@@ -9,15 +10,20 @@ pub mod seal;
 mod registry;
 mod types;
 
-pub use crate::registry::{RegisteredPoStProof, RegisteredSealProof, Version};
+pub use crate::registry::{RegisteredAggregationProof, RegisteredPoStProof, RegisteredSealProof};
 pub use crate::types::{PrivateReplicaInfo, PublicReplicaInfo};
 
-pub use filecoin_proofs_v1::storage_proofs::error::Error as StorageProofsError;
-pub use filecoin_proofs_v1::storage_proofs::fr32;
-pub use filecoin_proofs_v1::storage_proofs::post::election::Candidate;
-pub use filecoin_proofs_v1::storage_proofs::sector::{OrderedSectorSet, SectorId};
 pub use filecoin_proofs_v1::types::{
-    ChallengeSeed, Commitment, PaddedBytesAmount, PieceInfo, PoStType, ProverId, Ticket,
-    UnpaddedByteIndex, UnpaddedBytesAmount,
+    AggregateSnarkProof, ChallengeSeed, Commitment, PaddedBytesAmount, PieceInfo, PoStType,
+    ProverId, Ticket, UnpaddedByteIndex, UnpaddedBytesAmount,
 };
 pub use filecoin_proofs_v1::{FallbackPoStSectorProof, SnarkProof, VanillaProof};
+pub use fr32;
+pub use storage_proofs_core::{
+    api_version::ApiVersion,
+    error::Error as StorageProofsError,
+    merkle::MerkleTreeTrait,
+    parameter_cache::{get_parameter_data, get_verifying_key_data},
+    sector::{OrderedSectorSet, SectorId},
+    util::NODE_SIZE,
+};